Bors-NG v0.1.0 BorsNG.AdminController

Functionality that is specific to administrators.

Administrators are users with the is_admin flag set. Ensuring non-administrators can’t get here is done in BorsNG.Router.

Administrators have the ability to step outside the permissions and edit any repo, or delete it if they wish.

Functions

crashes(conn, map)

dup_patches(conn, params)

index(conn, params)

orphans(conn, params)

project_by_name(conn, map)

synchronize_all_installations(conn, params)