Bors-NG v0.1.0 BorsNG.Database.Batch

The database-level representation of a “batch”.

A batch is a collection of patches that are running, or will run.

Functions

all_assoc()

all_assoc(atom)

all_for_patch(patch_id, state \\ nil)

all_for_project(project_id)

all_for_project(project_id, state)

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

get_assoc_by_commit(project_id, commit, state \\ nil)

get_next_poll_unix_sec(batch)

is_empty(batch_id, repo)

new(project_id, into_branch, priority \\ 0)

next_poll_is_past(batch)

next_poll_is_past(batch, now_utc_sec)

timeout_is_past(batch)