Bors-NG v0.1.0 BorsNG.WobserverSecurity

An implementation of the wobserver security policy. This is necessary for one reason: Bors-NG runs on Heroku, which establishes every connection from a different load balancer IP. The default security policy rejects connections from a different IP than

Functions

authenticate(conn)

authenticate(Plug.Conn.t()) :: Plug.Conn.t()

authenticated?(conn)

authenticated?(Plug.Conn.t()) :: boolean()

authenticated?(:cowboy_req.req()) :: boolean()