Bors-NG v0.1.0 BorsNG.Worker.Syncer

A background task that pulls a full list of opened pull requests from a repo. Patches that don’t come up get closed, and patches that don’t exist get created.

Types

tcollaborator()

tcollaborator() :: BorsNG.GitHub.tcollaborator()

trepo_perm()

trepo_perm() :: BorsNG.GitHub.trepo_perm()

Functions

do_synchronize!(project_id, arg)

start_synchronize_project(project)

sync_patch(project_id, pr)

sync_patch(integer(), BorsNG.GitHub.Pr.t()) :: BorsNG.Database.Patch.t()

sync_user(gh_user)

sync_user(BorsNG.GitHub.User.t()) :: %BorsNG.Database.User{
  __meta__: term(),
  id: term(),
  inserted_at: term(),
  is_admin: term(),
  login: term(),
  projects: term(),
  updated_at: term(),
  user_xref: term()
}

synchronize_patches(open_patches, open_prs)

synchronize_patches(
  [
    %BorsNG.Database.Patch{
      __meta__: term(),
      author: term(),
      author_id: term(),
      body: term(),
      commit: term(),
      id: term(),
      inserted_at: term(),
      into_branch: term(),
      open: term(),
      pr_xref: term(),
      priority: term(),
      project: term(),
      project_id: term(),
      title: term(),
      updated_at: term()
    }
  ],
  [
    %BorsNG.GitHub.Pr{
      base_ref: term(),
      base_repo_id: term(),
      body: term(),
      head_ref: term(),
      head_repo_id: term(),
      head_sha: term(),
      merged: term(),
      number: term(),
      state: term(),
      title: term(),
      user: term()
    }
  ]
) :: [
  open: %BorsNG.GitHub.Pr{
    base_ref: term(),
    base_repo_id: term(),
    body: term(),
    head_ref: term(),
    head_repo_id: term(),
    head_sha: term(),
    merged: term(),
    number: term(),
    state: term(),
    title: term(),
    user: term()
  },
  close: %BorsNG.Database.Patch{
    __meta__: term(),
    author: term(),
    author_id: term(),
    body: term(),
    commit: term(),
    id: term(),
    inserted_at: term(),
    into_branch: term(),
    open: term(),
    pr_xref: term(),
    priority: term(),
    project: term(),
    project_id: term(),
    title: term(),
    updated_at: term()
  }
]

Returns a list of all patches that should be synchronized.

Note: This will return a list of all opened pull requests, as well as a list of patches that should be closed. It does not perform any filtering on open PRs, because those still need to have their metadata synced.

synchronize_project(project_id)

synchronize_project_collaborators(repo_conn, project_id)

synchronize_project_collaborators_by_role(project, collaborators, association, github_perm)

synchronize_project_collaborators_by_role(
  BorsNG.Database.Project.t(),
  [tcollaborator()],
  :users | :members,
  trepo_perm() | nil
) :: {:ok, BorsNG.Database.Project.t()} | {:error, any()}

wait_hot_spin(project_id)

Wait for synchronization to finish by hot-spinning. Used in test cases.